a lot of ui fixing

+++ iChronoMe.Android/Source/BaseActivity.cs
@@ -606,7 +606,7 @@ namespace iChronoMe.Droid
                 textView.Text += ex.Message;
                 sys.LogException(ex);
             }
-            int pad = 5 * sys.DisplayDensity;
+            int pad = (int)(5 * sys.DisplayDensity);
             textView.SetPadding(pad, pad, pad, pad);
             textView.MovementMethod = LinkMovementMethod.Instance;
             var scroll = new ScrollView(this);
